requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
SQLAlchemy
snowflake-connector-python[pandas]
cryptography==41.0.7
slowapi
redis
fastapi-limiter
email-validator
pytz
APScheduler==3.10.4